protect routes in Laravel

+++ routes/web.php
@@ -52,13 +52,13 @@ Route::get('/register', function() {
     ]);
 })->name('register');
 
-Route::get('/clients', 'App\Http\Controllers\ClientController@index')->name('clients.index');
-Route::post('/clients', 'App\Http\Controllers\ClientController@store')->name('clients.store');
-Route::get('/clients/{id}', 'App\Http\Controllers\ClientController@show')->name('clients.show');
-Route::get('/clients-create', 'App\Http\Controllers\ClientController@create')->name('clients.create');;
-Route::delete('/clients/{id}', 'App\Http\Controllers\ClientController@destroy')->name('clients.destroy');;
+Route::get('/clients', 'App\Http\Controllers\ClientController@index')->name('clients.index')->middleware('auth');
+Route::post('/clients', 'App\Http\Controllers\ClientController@store')->name('clients.store')->middleware('auth');
+Route::get('/clients/{id}', 'App\Http\Controllers\ClientController@show')->name('clients.show')->middleware('auth');
+Route::get('/clients-create', 'App\Http\Controllers\ClientController@create')->name('clients.create')->middleware('auth');
+Route::delete('/clients/{id}', 'App\Http\Controllers\ClientController@destroy')->name('clients.destroy')->middleware('auth');
 
-Route::get('/clients/code/{code}', 'App\Http\Controllers\ClientController@where');
+Route::get('/clients/code/{code}', 'App\Http\Controllers\ClientController@where')->middleware('auth');;
 
 Auth::routes();
